requirements.txt
pillow==10.3.0
pyheif==0.7.1
